LayoutTests/imported/w3c/web-platform-tests/fetch/compression-dictionary/dictionary-decompression.tentative.https-expected.txt


FAIL Decompresion using Brotli with the dictionary works as expected assert_equals: expected ":U5abz16WDg7b8KS93msLPpOB4Vbef1uRzoORYkJw9BY=:" but got "\"available-dictionary\" header is not available"
FAIL Decompresion using Zstandard with the dictionary works as expected assert_equals: expected ":U5abz16WDg7b8KS93msLPpOB4Vbef1uRzoORYkJw9BY=:" but got "\"available-dictionary\" header is not available"
FAIL Decompresion of a cross origin resource works as expected assert_equals: expected ":U5abz16WDg7b8KS93msLPpOB4Vbef1uRzoORYkJw9BY=:" but got "\"available-dictionary\" header is not available"
